+++ spec/spec_helper.rb
@@ -33,10 +33,22 @@ module HandBrake
     # A stub implementation of the runner for HandBrake::CLI.
     class StaticRunner
       attr_accessor :output, :status
+      ##
+      # A lambda that specifies desired side effects of execution or
+      # simulates things happening outside the execution but
+      # simultaneous with it.
+      attr_accessor :behavior
       attr_reader :actual_arguments
 
       def run(args)
         @actual_arguments = args
+        if i = args.index('--output')
+          fn = args[i + 1]
+          File.open(fn, 'w') { |f| f.write 'This is the file created by --output' }
+        end
+        if behavior
+          behavior.call
+        end
         HandBrake::CLI::RunnerResult.new(output, status)
       end