+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Support overwrite detection and behaviors for the
   {HandBrake::CLI#output}.
+- Support "atomic" output mode in {HandBrake::CLI#output}.
 
 0.1.0
 =====